+++ src/Recyclarr.Core/Config/Parsing/ConfigYamlDataObjects.cs
@@ -19,7 +19,6 @@ public record CustomFormatConfigYaml
 {
     public IReadOnlyCollection<string>? TrashIds { get; init; }
     public IReadOnlyCollection<QualityScoreConfigYaml>? AssignScoresTo { get; init; }
-    public IReadOnlyCollection<QualityScoreConfigYaml>? QualityProfiles { get; init; }
 }
 
 [UsedImplicitly(ImplicitUseKindFlags.Assign, ImplicitUseTargetFlags.WithMembers)]