+++ src/Recyclarr.Core/Config/Parsing/ErrorHandling/ConfigContextualMessages.cs
@@ -28,6 +28,12 @@ public static class ConfigContextualMessages
                 + "See: <https://recyclarr.dev/guide/upgrade-guide/v8.0/#replace-existing-removed>";
         }
 
+        if (msg.Contains("Property 'quality_profiles' not found on type", StringComparison.Ordinal))
+        {
+            return "The `quality_profiles` element under `custom_formats` has been renamed to `assign_scores_to`. "
+                + "See: <https://recyclarr.dev/guide/upgrade-guide/v8.0/#assign-scores-to>";
+        }
+
         return null;
     }
 }